break out loader into its own component
