+++ src/site/src/components/siteplotly.js
@@ -1,8 +1,6 @@
 import React from "react"
 import Loadable from "react-loadable"
-import Loader from "react-loader-spinner"
-
-const spinnerStyle = { textAlign: "center" }
+import Loading from "./loading"
 
 const Plotly = Loadable({
   loader: () => import(`react-plotly.js`),
@@ -10,15 +8,7 @@ const Plotly = Loadable({
     timedOut ? (
       <blockquote>Error: Loading Plotly timed out.</blockquote>
     ) : (
-      <div style={spinnerStyle}>
-        <Loader
-          type="ThreeDots"
-          color="#525453"
-          height={75}
-          width={75}
-          timeout={10000} // 10sec
-        />
-      </div>
+      <Loading />
     ),
   timeout: 10000,
 })
